Index

QuantumStateBase.ComplexVec

QuantumStateBase.Annihilation

QuantumStateBase.CoherentState

QuantumStateBase.Creation

QuantumStateBase.Displacement

QuantumStateBase.FockState

QuantumStateBase.SinglePhotonState

QuantumStateBase.SqueezedState

QuantumStateBase.SqueezedThermalState

QuantumStateBase.Squeezing

QuantumStateBase.ThermalState

QuantumStateBase.VacuumState

QuantumStateBase.annihilate

QuantumStateBase.create

QuantumStateBase.displace

QuantumStateBase.laguerre

QuantumStateBase.laguerre

QuantumStateBase.squeeze

QuantumStateBase.z

APIs

QuantumStateBase.ComplexVec — Method

QuantumStateBase.ComplexVec(r::T, θ::T)

Vector in polar coordinate for complex plane.

$v = re^{-i\theta}$

Arguments

r: Radius of the complex number.

θ: Argument of the complex number.

QuantumStateBase.Annihilation — Method

QuantumStateBase.Annihilation(T; dim::Integer)

Annihilation operator in matrix form.

$\hat{a}$

QuantumStateBase.CoherentState — Method

CoherentState([T::Type{<:Complex}], r::Real, θ::Real, [rep::Type]; dim::Integer)

Construct a coherent state in rep representation.

T: Numerical type. Default as ComplexF64 if not assigned.

r: The radius in phase space.

θ: The phase angle in phase space.

rep:In which representation, default as Vector if not assigned.

dim: Maximum photon number for truncate.

QuantumStateBase.Creation — Method

QuantumStateBase.Creation(T; dim::Integer)

Creation operator in matrix form.

$\hat{a}^{\dagger}$

QuantumStateBase.Displacement — Method

QuantumStateBase.Displacement(T::Type{<:Complex}, r::Real, θ::Real; dim::Integer)

Displacement operator in matrix form.

$\hat{D}(\alpha) = exp(\alpha \hat{a}^{\dagger} - \alpha^{*} \hat{a})$

QuantumStateBase.FockState — Method

FockState([T::Type{<:Number}], n::Integer, [rep::Type]; dim::Integer)

Construct a Fock state in rep representation.

Arguments

T: Numerical type. Default as Float64 if not assigned.

n: Photon number of Fock state.

rep: In which representation, default as Vector if not assigned.

dim: Maximum photon number for truncate.

QuantumStateBase.SinglePhotonState — Method

SinglePhotonState([T::Type{<:Number}], [rep::Type]; dim::Integer)

Construct a single photon state in rep representation.

Arguments

T: Numerical type. Default as Float64 if not assigned.

rep:In which representation, default as Vector if not assigned.

dim: Maximum photon number for truncate.

QuantumStateBase.SqueezedState — Method

SqueezedState([T::Type{<:Complex}], r::Real, θ::Real, [rep::Type]; dim::Integer)

Squeezed state is defined if its electric field strength for some phases has a quantum uncertainty smaller than that of a coherent state.

Arguments

T: Numerical type. Default as ComplexF64 if not assigned.

r: The squeezing level.

θ: The squeezing phase.

rep:In which representation, default as Vector if not assigned.

dim: Maximum photon number for truncate.

This constructor will construct $| \xi \rangle = \hat{S}(\xi) | 0 \rangle$

QuantumStateBase.SqueezedThermalState — Method

SqueezedThermalState([T::Type{<:Complex}], r::Real, θ::Real, n̄::Real; dim::Integer)

Squeezed thermal state is defined if its electric field strength for some phases has a quantum uncertainty smaller than that of a coherent state.

Arguments

T: Numeric type. Default as ComplexF64 if not assigned.

r: Squeezing level.

θ: Squeezing phase.

n̄: Average photon number at a certain temperature.

dim: Maximum photon number for truncate.

This constructor will construct $\rho = \hat{S}(\xi) \rho_{th} \hat{S}(\xi)^{T}$

QuantumStateBase.Squeezing — Method

QuantumStateBase.Squeezing(T::Type{<:Complex}, r::Real, θ::Real; dim::Integer)

Squeezing operator in matrix form.

Arguments

r: Squeezing level.

θ: Squeezing phase.

$\hat{S}(\xi) = exp(\frac{1}{2}(\xi^{*} \hat{a}^{2} - \xi \hat{a}^{\dagger 2}))$

QuantumStateBase.ThermalState — Method

ThermalState([T::Type{<:Number}], n̄::Real; dim::Integer)

Thermal state is a mixed state with photon number distribution described by Bose-Einstein distribution.

Arguments

T: Numerical type. Default as Float64 if not assigned.

n̄: Average photon number at a certain temperature.

dim: Maximum photon number for truncate.

QuantumStateBase.VacuumState — Method

VacuumState([T::Type{<:Number}], [rep::Type]; dim::Integer)

Construct a vacuum state in rep representation.

Arguments

T: Numerical type. Default as Float64 if not assigned.

rep:In which representation, default as Vector if not assigned.

dim: Maximum photon number for truncate.

QuantumStateBase.annihilate — Method

annihilate(s::AbstractArray{T})

Apply annihilation operator on a quantum state.

Arguments

s: s can be a quantum state vector or a density matrix of a quantum state.

Examples

julia> state = SinglePhotonState(dim=5);

julia> new_state = annihilate(state);

julia> new_state == VacuumState(dim=5)
true

julia> state = SinglePhotonState(Matrix, dim=5);

julia> new_state = annihilate(state);

julia> new_state == VacuumState(Matrix, dim=5)
true

QuantumStateBase.create — Method

create(s::AbstractArray{T})

Apply creation operator on a quantum state.

Arguments

s: s can be a quantum state vector or a density matrix of a quantum state.

Examples

julia> state = VacuumState(dim=5);

julia> new_state = create(state);

julia> new_state == SinglePhotonState(dim=5)
true

julia> state = VacuumState(Matrix, dim=5);

julia> new_state = create(state);

julia> new_state == SinglePhotonState(Matrix, dim=5)
true

QuantumStateBase.displace — Method

displace(s::AbstractArray{T}, r, θ)

Apply displacement operator on a quantum state.

Arguments

s: s can be a quantum state vector or a density matrix of a quantum state.

r: Radius of displacement.

θ: Phase of displacement.

Examples

julia> state = VacuumState(dim=5);

julia> new_state = displace(state, 2, π/4);

julia> new_state == CoherentState(2, π/4, dim=5)
true

QuantumStateBase.laguerre — Method

laguerre(n::Integer, α::Integer)(x::Real)

Generalized Laguerre polynomials

Examples

julia> L = laguerre(1, 3);

julia> L(5)
-1

QuantumStateBase.laguerre — Method

laguerre(n::Integer, α::Integer, x::Real)

Generalized Laguerre polynomials

Examples

julia> laguerre(0, 1, 5)
1

julia> laguerre(1, 3, 5)
-1

QuantumStateBase.squeeze — Method

squeeze(s::AbstractArray{T})

Apply squeezing operator on a quantum state.

Arguments

s: s can be a quantum state vector or a density matrix of a quantum state.

r: Squeezing level.

θ: Squeezing phase.

Examples

julia> state = VacuumState(dim=5);

julia> new_state = squeeze(state, 0.5, π/4);

julia> new_state == SqueezedState(0.5, π/4, dim=5)
true

julia> state = VacuumState(Matrix, dim=5);

julia> new_state = squeeze(state, 0.5, π/4);

julia> new_state == SqueezedState(0.5, π/4, Matrix, dim=5)
true

QuantumStateBase.z — Method

QuantumStateBase.z(complexvec::ComplexVec)

Evaluate the complex vector.