Index

QuantumStateBase.ComplexVec

QuantumStateBase.WignerFunction

QuantumStateBase.Annihilation

QuantumStateBase.CoherentState

QuantumStateBase.Creation

QuantumStateBase.Displacement

QuantumStateBase.FockState

QuantumStateBase.SinglePhotonState

QuantumStateBase.SqueezedState

QuantumStateBase.SqueezedThermalState

QuantumStateBase.Squeezing

QuantumStateBase.ThermalState

QuantumStateBase.VacuumState

QuantumStateBase.annihilate

QuantumStateBase.create

QuantumStateBase.displace

QuantumStateBase.laguerre

QuantumStateBase.laguerre

QuantumStateBase.squeeze

QuantumStateBase.wigner

QuantumStateBase.z

APIs

QuantumStateBase.ComplexVec — Method

QuantumStateBase.ComplexVec(r::T, θ::T)

Construct a vector in polar coordinate for complex plane.

$v = re^{-i\theta}$

Arguments

r: Radius of the complex number.

θ: Argument of the complex number.

QuantumStateBase.WignerFunction — Type

WignerFunction

WignerFunction to calculate Wigner function for a quantum state in gevin x and p range.

Arguments

x_range: First physical domain range in phase space.

p_range: Second physical domain range in phase space.

dim: Photon number truncation of the quantum state.

Example

julia> xs = -1:0.1:1; ps = -1:0.1:1; dim=35;

julia> wf = WignerFunction(xs, ps, dim=dim);

julia> wf(VacuumState(Matrix, dim=dim));

QuantumStateBase.Annihilation — Method

QuantumStateBase.Annihilation(T; dim::Integer)

Construct an annihilation operator in matrix form.

$\hat{a}$

QuantumStateBase.CoherentState — Method

CoherentState([T::Type{<:Complex}], r::Real, θ::Real, [rep::Type]; dim::Integer)

Construct a coherent state in rep representation.

T: Numerical type. Default as ComplexF64 if not assigned.

r: The radius in phase space.

θ: The phase angle in phase space.

rep:In which representation, default as Vector if not assigned.

dim: Maximum photon number for truncate.

QuantumStateBase.Creation — Method

QuantumStateBase.Creation(T; dim::Integer)

Construct a creation operator in matrix form.

$\hat{a}^{\dagger}$

QuantumStateBase.Displacement — Method

QuantumStateBase.Displacement(T::Type{<:Complex}, r::Real, θ::Real; dim::Integer)

Construct a displacement operator in matrix form.

$\hat{D}(\alpha) = exp(\alpha \hat{a}^{\dagger} - \alpha^{*} \hat{a})$

QuantumStateBase.FockState — Method

FockState([T::Type{<:Number}], n::Integer, [rep::Type]; dim::Integer)

Construct a Fock state in rep representation.

Arguments

T: Numerical type. Default as Float64 if not assigned.

n: Photon number of Fock state.

rep: In which representation, default as Vector if not assigned.

dim: Maximum photon number for truncate.

QuantumStateBase.SinglePhotonState — Method

SinglePhotonState([T::Type{<:Number}], [rep::Type]; dim::Integer)

Construct a single photon state in rep representation.

Arguments

T: Numerical type. Default as Float64 if not assigned.

rep:In which representation, default as Vector if not assigned.

dim: Maximum photon number for truncate.

QuantumStateBase.SqueezedState — Method

SqueezedState([T::Type{<:Complex}], r::Real, θ::Real, [rep::Type]; dim::Integer)

Construct a squeezed state in rep representation.

Squeezed state is defined if its electric field strength for some phases has a quantum uncertainty smaller than that of a coherent state.

Arguments

T: Numerical type. Default as ComplexF64 if not assigned.

r: The squeezing level.

θ: The squeezing phase.

rep:In which representation, default as Vector if not assigned.

dim: Maximum photon number for truncate.

This constructor will construct $| \xi \rangle = \hat{S}(\xi) | 0 \rangle$

QuantumStateBase.SqueezedThermalState — Method

SqueezedThermalState([T::Type{<:Complex}], r::Real, θ::Real, n̄::Real; dim::Integer)

Construct a squeezed thermal state.

Squeezed thermal state is defined if its electric field strength for some phases has a quantum uncertainty smaller than that of a coherent state.

Arguments

T: Numeric type. Default as ComplexF64 if not assigned.

r: Squeezing level.

θ: Squeezing phase.

n̄: Average photon number at a certain temperature.

dim: Maximum photon number for truncate.

This constructor will construct $\rho = \hat{S}(\xi) \rho_{th} \hat{S}(\xi)^{T}$

QuantumStateBase.Squeezing — Method

QuantumStateBase.Squeezing(T::Type{<:Complex}, r::Real, θ::Real; dim::Integer)

Construct a squeezing operator in matrix form.

Arguments

r: Squeezing level.

θ: Squeezing phase.

$\hat{S}(\xi) = exp(\frac{1}{2}(\xi^{*} \hat{a}^{2} - \xi \hat{a}^{\dagger 2}))$

QuantumStateBase.ThermalState — Method

ThermalState([T::Type{<:Number}], n̄::Real; dim::Integer)

Construct a thermal state.

Thermal state is a mixed state with photon number distribution described by Bose-Einstein distribution.

Arguments

T: Numerical type. Default as Float64 if not assigned.

n̄: Average photon number at a certain temperature.

dim: Maximum photon number for truncate.

QuantumStateBase.VacuumState — Method

VacuumState([T::Type{<:Number}], [rep::Type]; dim::Integer)

Construct a vacuum state in rep representation.

Arguments

T: Numerical type. Default as Float64 if not assigned.

rep:In which representation, default as Vector if not assigned.

dim: Maximum photon number for truncate.

QuantumStateBase.annihilate — Method

annihilate(s::AbstractArray{T})

Apply annihilation operator on a quantum state.

Arguments

s: s can be a quantum state vector or a density matrix of a quantum state.

Examples

julia> state = SinglePhotonState(dim=5)
5-element Vector{Float64}:
 0.0
 1.0
 0.0
 0.0
 0.0

julia> annihilate(state)
5-element Vector{Float64}:
 1.0
 0.0
 0.0
 0.0
 0.0

julia> state = SinglePhotonState(Matrix, dim=5)
5×5 Matrix{Float64}:
 0.0  0.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0

julia> annihilate(state)
5×5 Matrix{Float64}:
 1.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0

QuantumStateBase.create — Method

create(s::AbstractArray{T})

Apply creation operator on a quantum state.

Arguments

s: s can be a quantum state vector or a density matrix of a quantum state.

Examples

julia> state = VacuumState(dim=5)
5-element Vector{Float64}:
 1.0
 0.0
 0.0
 0.0
 0.0

julia> create(state)
5-element Vector{Float64}:
 0.0
 1.0
 0.0
 0.0
 0.0

julia> state = VacuumState(Matrix, dim=5)
5×5 Matrix{Float64}:
 1.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0

julia> create(state)
5×5 Matrix{Float64}:
 0.0  0.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0

QuantumStateBase.displace — Method

displace(s::AbstractArray{T}, r, θ)

Apply displacement operator on a quantum state.

Arguments

s: s can be a quantum state vector or a density matrix of a quantum state.

r: Radius of displacement.

θ: Phase of displacement.

Examples

julia> state = VacuumState(dim=5)
5-element Vector{Float64}:
 1.0
 0.0
 0.0
 0.0
 0.0

julia> displace(state, 2, π/4)
5-element Vector{ComplexF64}:
   0.14864196494651452 - 5.551115123125783e-17im
   0.15309939720606983 - 0.15309939720606955im
 8.326672684688674e-17 - 0.5201876236174409im
   -0.1359447277638306 - 0.13594472776383068im
   -0.7896009204837534 - 3.0531133177191805e-16im

julia> state = VacuumState(Matrix, dim=5)
5×5 Matrix{Float64}:
 1.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0

julia> displace(state, 2, π/4)
5×5 Matrix{ComplexF64}:
   0.0220944+0.0im              0.022757+0.022757im   4.12531e-17+0.0773217im    -0.0202071+0.0202071im    -0.117368+8.92137e-17im
    0.022757-0.022757im        0.0468789+0.0im          0.0796404+0.0796404im  -2.42861e-17+0.0416261im    -0.120887+0.120887im
 4.12531e-17-0.0773217im       0.0796404-0.0796404im     0.270595+0.0im           0.0707168+0.0707168im  9.30717e-17+0.410741im
  -0.0202071-0.0202071im    -2.42861e-17-0.0416261im    0.0707168-0.0707168im     0.0369619+0.0im           0.107342+0.107342im
   -0.117368-8.92137e-17im     -0.120887-0.120887im   9.30717e-17-0.410741im       0.107342-0.107342im       0.62347+0.0im

QuantumStateBase.laguerre — Method

laguerre(n::Integer, α::Integer)(x::Real)

Generalized Laguerre polynomials

Examples

julia> L = laguerre(1, 3);

julia> L(5)
-1

QuantumStateBase.laguerre — Method

laguerre(n::Integer, α::Integer, x::Real)

Generalized Laguerre polynomials

Examples

julia> laguerre(0, 1, 5)
1

julia> laguerre(1, 3, 5)
-1

QuantumStateBase.squeeze — Method

squeeze(s::AbstractArray{T})

Apply squeezing operator on a quantum state.

Arguments

s: s can be a quantum state vector or a density matrix of a quantum state.

r: Squeezing level.

θ: Squeezing phase.

Examples

julia> state = VacuumState(dim=5)
5-element Vector{Float64}:
 1.0
 0.0
 0.0
 0.0
 0.0

julia> squeeze(state, 0.5, π/4)
5-element Vector{ComplexF64}:
     0.9419264274916698 + 0.0im
                    0.0 + 0.0im
   -0.21510388914370943 + 0.21510388914370937im
                    0.0 + 0.0im
 3.0228557826221164e-17 - 0.1422506201859301im

julia> state = VacuumState(Matrix, dim=5)
5×5 Matrix{Float64}:
 1.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0

julia> squeeze(state, 0.5, π/4)
5×5 Matrix{ComplexF64}:
    0.887225+0.0im       0.0+0.0im   -0.202612-0.202612im   0.0+0.0im  2.84731e-17+0.13399im
         0.0+0.0im       0.0+0.0im         0.0+0.0im        0.0+0.0im          0.0+0.0im
   -0.202612+0.202612im  0.0+0.0im   0.0925394+0.0im        0.0+0.0im   -0.0305987-0.0305987im
         0.0+0.0im       0.0+0.0im         0.0+0.0im        0.0+0.0im          0.0+0.0im
 2.84731e-17-0.13399im   0.0+0.0im  -0.0305987+0.0305987im  0.0+0.0im    0.0202352+0.0im

QuantumStateBase.wigner — Method

wigner(ρ::AbstractMatrix{T}, x_range::AbstractRange, p_range::AbstractRange)

Calculate the Wigner fuction of the quantum state in the gevin x and p range.

Arguments

ρ: Density matrix of a quantum state.

x_range: First physical domain range in phase space.

p_range: Second physical domain range in phase space.

Example

julia> xs = -1:0.1:1; ps = -1:0.1:1;

julia> wigner(VacuumState(Matrix, dim=35), xs, ps);

QuantumStateBase.z — Method

QuantumStateBase.z(complexvec::ComplexVec)

Evaluate the complex vector.